user books fetching

+++ src/controller/books.controller.js
@@ -83,6 +83,10 @@ const deleteBook = asyncHandler(async (req, res) => {
         throw new ApiError(404, "Book Not Found");
     }
 
+    if (book.owner.toString() !== req.user._id.toString()) {
+        throw new ApiError(403, "You are not authorized to delete this Book")
+    }
+
     const deletedThumbnail = await deleteImageFromCloudinary(book.thumbnail);
     if (!deletedThumbnail) {
         throw new ApiError(500, "Thumbnail is not deleted");
@@ -107,6 +111,10 @@ const updateBook = asyncHandler(async (req, res) => {
     const book = await Books.findById(bookId);
     if (!book) {
         throw new ApiError(404, "Book Not Found");
+    }
+
+    if (book.owner.toString() !== req.user._id.toString()) {
+        throw new ApiError(403, "You are not authorized to update this Book")
     }
 
     const updatedBook = await Books.findByIdAndUpdate(
@@ -134,6 +142,10 @@ const updateBookThumbnail = asyncHandler(async (req, res) => {
     const book = await Books.findById(bookId);
     if (!book) {
         throw new ApiError(404, "Book not found");
+    }
+
+    if (book.owner.toString() !== req.user._id.toString()) {
+        throw new ApiError(403, "You are not authorized to update this Book Thumbnail")
     }
 
     if (book.thumbnail) {
